refactor: seperate bookmarked tab into standalone component
